speed-up Inflate.evalf
This patch removes assertions in `Inflate.evalf` that are costly compared to
`numpy.add.at` if the size of `indices` and `array` is very small. With
`NUTILS_DEBUG=evalf` the same checks are performed before calling
`Inflate.evalf`.

In addition this patch stores the tuple of slices for the pointwise axes in an
attribute, which increases the performance of `Inflate.evalf` a little bit.