add inplace add evaluable
The sum of several sparse `Inflate`s is costly because the sum operates on the
dense, inflated arrays. This patch aims to reduce the cost by introducing
inplace additions. First there is the notion of `Array.evalf_iadd(output,
*args)`, which, compared to `Array.evalf(*args)`, receives the output array as
argument. Implementations must add the result to the output instead of
returning the result. Calling `Array.evalf_iadd` is left for special-purpose
`Evaluable`s which gobble `Arrays` and their arguments and call the
`evalf_iadd`s inside it's `evalf`. This patch introduces one such `Evaluable`:
`_OfnAddInplace`.