fix failing tests using parallel evaluable.compile
When replacing `Evaluable.eval` with a parallel `evaluable.compile` several
assertions in the caching mechanism of `solver` and corresponding unittests
fail because a repeated call to `evaluable.compile` may produce slightly
different results. This is caused by the non-deterministic order of inplace
adds in the different threads. This patch removes assertions and rewrites
unittests using tolerances.