assemble into sorted and deduped coo directly
Currently we assemble sparse matrices as unstructured coo with repeated indices
and sort and dedup later. The downside of this approach is that we're wasting
memory because of the duplicated indices, and if we're assembling a matrix more
than once, e.g. for Newton, we always have to sort and dedup, even if the
sparse structure doesn't change.

This patch changes `Evaluable.assparse` such that the indices are sorted and
deduped first, and the values are then inflated using the unique indices.