remove shape[:-1] from LoopConcat constructor args
The shape of `LoopConcatenate` is used in the loop initialization to allocate
the output array. The length of the concatenation axis is the result of
concatenating and summing (`_SizesToOffset`) the lengths of the same axis of
the concatenation value. Under certain conditions the concatenation length
can be simplified. For this to work, the concatenation length must be a
constructor argument of `LoopConcatenate`, otherwise the length is not picked
up by `deep_replace_property`.

For the same reason the entire shape of `LoopConcatenate` was passed as
constructor argument. Since the shape of an array is very likely already
optimized --- the same is true for `LoopConcatenate.shape`! --- this patch
removes the shape from the constructor of `LoopConcatenate`, except for the
concatenation axis.