repl ArrayFromTuple(_lower, _upper) w tuple attr
Computing the integer bounds can be expensive. For example, a scalar array that
does not depend on arguments will be simplified and evaluated, regardless how
expensive this evaluation is and disregarding the `_intbounds_impl`.

Currently the `ArrayFromTuple` class takes integer bounds as arguments. When
combining loops, the integer bounds for the loops are evaluated, the loops are
combined and the evaluated integer bounds are passed to `ArrayFromTuple`, even
if we are not using the bounds.

To prevent evaluating the bounds, this patch introduces the `_intbounds_tuple`
property for `Evaluable`s that return tuples and replaces the integer bounds
argument from `ArrayFromTuple` with a `_intbounds_impl` that queries the
`_intbounds_tuple`.