replace loop name with id object
To facilitate modifying the name of a `Loop` and the correspondong `_LoopIndex`
with `util.shallow_replace`, this patch replaces the loop name, a `str`, with a
`_LoopId` object wrapping the name. The new `_LoopId` supports not only `str`s,
but any immutable object as name or identifier. The use of non-`str` ids is
reserved for interal use, hence `loop_index()`, the public constructor of
`_LoopIndex`, now tests if the name is a `str`.