handle trivial transposes directly
This patch makes the trans argument of evaluable.transpose mandatory and adds a
simplification path for the trivial case. As transpose is used in the pre- and
post processing of almost all other operations there are considerable merits to
handling this case directly. To avoid double work, the Transpose operation no
longer accepts the trivial permutation as an argument.