support numpy.number array bases in lru_cache
`types.lru_cache` uncovers the bases of `numpy.ndarray` arguments and stores
weak references to the bases with a clean-up function. If the base is a
`numpy.number`, then it is impossible to get a weak reference and `lru_cache`
raises. This patch fixes the problem by holding a weak reference to the array
that references the base if the base is a `numpy.number`.

The above situation occurs when calling `numpy.ndarray` with a `numpy.number`
as base, which is used in `evaluable.InsertAxis.evalf`.