add evaluable.compile: a Python code generator
This patch adds an alternative for `Evaluable.eval`: an Evaluable to Python
code generator. This patch adds a new module `nutils._pyast` which provides a
small subset of the Python language in the form of AST nodes, and a function
`evaluable.compile` which compiles `Evaluable`s to AST, translates AST to
Python code and returns the compiled Python code as a callable.