alt mode for Evaluable._node with unique loop ids
If the loop ids are unique, it is unnecessary for the `times` parameter of
`Evaluable._node` to have nested dicts at `Loop`s. This patch adds a
`unique_loop_ids` parameter to `Evaluable._node` which, if true, indicates that
the `times` dict does not contain nested dicts and can be passed directly to
the loop body.