add evaluable._make_loop_ids_unique
Working with `evaluable.Loop`s can be difficult if loop ids are reused in
different loops. Consider for example the following tree:

    A: LoopSum i
      B: Add
        C: LoopSum i
          D#C: _LoopIndex i
        D#A

Both loops A and C use the same id i. Consequently it is impossible to relate
D#A (the hash suffix is added for clarifation) to A and D#C to C other than by
inspecting the tree between the loop and the index: some index I belongs to
loop L if there is no other loop in between that uses the same id.

To make life easier, this patch adds `evaluable._make_loop_ids_unique`, which
replaces all loop ids of target funcs with unique loop ids if there are
duplicate ids.