add _util.untake
This patch adds the _util.untake helper method that can be used as an
alternative to numpy.argsort in case the indices are a permuted range, with the
benefit that the resulting indices are Python integers. Moreover, a second
argument can be provided to shuffle an interable rather than returning indices.