Add matrix.assemble_coo, matrix.assemble_csr
This patch adds matrix.assemble_csr and matrix.assemble_coo, where the
latter is the same in nature as the existing matrix.assemble function
but with a different signature and implementation.

The assemble function of the backends are changed from COO to CSR input, since
this is closer to the internal representation of most sparse matrix backends.
Moreover, CSR data will be generated directly by the evaluable module in
future, so the conversion to COO and back would represent additional work.

The existing matrix.assemble function is maintained for backwards
compatibility, but will be deprecated in a subsequent commit (so we can
still test its validity here before changing all the unit tests) and
removed in Nutils 10.