Reimplement legacy solvers on top of System
This patch reimplements solver.solve_linear, solver.newton, solver.minimize,
solver.pseudotime, and solver.thetamethod by calling the corresponding methods
of solver.System. Lost in this transition are the ability to cache individual
newton and minimize iterations, to use droptol in nonlinear problems (which was
a problematic concept) and to use a non-constant interatia matrix in pseudotime.