Replace _equals_scalar_constant by isunit
This patch removes the evaluable._equals_scalar_constant function, which
was used only to test equality with the unit value, and replaces it with
a dedicated function evaluable.isunit analogous to evaluable.iszero. The
new function avoids evaluation; instead it tests if the simplified
evaluable is of type Constant with value 1.