Reduce _intbounds evaluations
This patch changes the logic in Array._intbounds with the aim to reduce
the number of potentially expensive array evaluations.

Formerly, evaluation always happened for scalar constant arrays. The
_intbounds_impl method was called only for non-scalar or non-constant
arrays.

The new implementation always calls _intbounds_impl, which evaluates
constant scalar arrays in its default implementation. The result is that
evaluation is avoided for all arrays that provide their own
_intbounds_impl.