Fix deprecated tol argument in scipy>=1.12
Scipy 1.12 (Jan 2024) deprecated the tol argument in its various solvers by
rtol, and scipy 1.14 (Jun 2024) removed the old syntax. Since these versions
are less than a year old, this patch introduces a check to provide the correct
argument based on the installed version, so as to remain compatible with older
versions of Scipy.