Add Array.assparse, evaluable.eval_coo
This patch adds the Array.assparse method which flattens and
concatenates blocks from _assparse. It also adds the eval_coo function
which compiles and evaluates the assparse output. This makes eval_coo an
alternative to eval_sparse, which returns a nutils.sparse array object.
This function will be deprecated in future, and removed along with the
entire nutils.sparse infrastructure.