Add evaluable.AssertEqual
This patch adds the runtime check evaluable.AssertEqual, and uses it in the
return shape of Add, Multiply, Einsum and Pointwise. While shape
inconsistencies will result in exceptions during the evaluation of these
operations, there are situations where the shape itself enters a computation,
in which case an explicit runtime check is required to catch inconsistencies.