Fix NaN output of numpy.linalg.norm
This patch adds a positivity guard to the implementation of numpy.linalg.norm.
While the sum of squares is always positive, we wrap it in maximum(0, ..) to
guard against the situation that the function simplifies to a form in which
round-off errors may nudge it below zero (e.g. (a-b)^2 -> a^2 - 2 a b + b^2).