Cache entire evaluable for constant linear systems
This patch avoids a second evaluation of a linear system if the matrix, vector
and value are all constant. Formerly only the matrix was cached, but the
changes of commit b7cdff34 imply that the the entire evaluable may become
constant, in which case the residual vector and function value can be reused as
well. A new flag is introduced to differentiate between the situation that all
items are constant (`is_constant`) or only the matrix (`is_constant_matrix`).