Return existing constraints in System.optimize
This patch changes the behaviour of System.optimize but having it return, in
addition to the solution arrays for self.trials, the existing values of the
constraints parameter rather than arguments. This is to accommodate the main
application of optimize, which is to form constraints for subsequent
invocations of solve. The new behaviour allows for constraint stacking by
feeding existing constraints as the constraints argument of subsequent
optimizations.