Do not raise SolverError for faulty parameters
This patch changes some SolverError exceptions to ValueError when the reason of
the error is directly apparent from checking the parameters, leaving the use of
SolverError exclusively for errors that occur _during_ the solution process,
for instance due to convergence issues. This fixes System.step's time halving
logic, which spsecifically catches SolverError and MatrixError, that would
repeat forever if the cause of the issue is independent of the timestep.