Introduce assert_equal, assert_equal_tuple
This patch adds the functions evaluable.assert_equal and
evaluable.assert_equal_tuple, which instantiate AssertEqual only if the two
arguments are different. This is important because the shape attribute is not
in simplified form if it is generated by the evaluable, and therefore is prone
to stack up with potentially detrimental effects.