Switch row to column support in solve_constraints
This patch removes rowsupp in System.solve_constraints and replaces it with a
local implementation for column support. Aside from preparing for the imminent
deprecation of rowsupp, switching to a column-based support test is more
appropriate now that solve_constraints is no longer limited to symmetric
systems, as the columns relate to the trial dofs that are being constrained.
When in future row constraints will be better supported then the local test can
be extended with a row test for non-symmetric systems.