Replace System.trialshapes by argshapes
This patch replaces the trialshapes dictionary in System, which contains only
the shapes of the trial arguments, by the argshapes dictionary, which contains
all of the evaluable's arguments. The reason for the change is that the broader
set of arguments can be used for all sorts of argument checks prior to
initiating a solve.