Fix lowering of function.factor
This patch changes the implementation of function.factor, from using _Unlower
to being a Array class with its own dedicated lower method. This fixes the
lowering of a factored function that is incorporated in a function with a
different space, in which case non-trivial lower args are provided. The newly
added unit tests function.test_factor.test_lower_with_points asserts that this
situation is now handled correctly.