Recurse argument degrees via main method
This patch replaces any recursion of the _argument_degree method to an
underlying _argument_degree by a recursion via argument_degree (sans
underscore), in the interest of debugging.

The double hop was intentionally avoided where possible to reduce call stack
depth. However, this resulted in NotPolynomial exceptions being raised from
higher up the stack, where it is very hard to retrace the error to the
offending operation.