Enhance the argument degree of Power
The Power operation reports an argument degree based on that of its base
argument if the power is a positive integer constant. Avoiding potentially
expensive evaluations, the latter is tested by an instance check. However, this
approach misses constants that are wrapped in a cast operation such as
IntToFloat. The current patch strips the power argument of casts before
proceding with the constness test.