Fix Gramm-Schmidt breakdown in Arnoldi solver
This patch fixes a division by zero error that may occur in the _solve_arnoldi
routine in case the linear solver returns a poor result. The fix simply breaks
out of the loop if this situation occurs, in accordance with the standing logic
that iterations continue until the solution no longer improves. An atol or rtol
argument must be provided to the solver to detect if the yielded result is of
sufficient quality.