Fix use of linalg.solve with multiple rhs
Since Nutils 2, linalg.solve will only recognize a rhs argument to be a stack
of vectors if its dimension equals that of the matrix stack. Calling solve with
a stack of matrices [A1, A2, A2] and a stack of rhs vectors [b1, b2, b3] no
longer works; instead, we must provide the rhs as [[b1], [b2], [b3]] and remove
the singleton axis from the result. This patch makes the required change to
Orthonormal.evalf and some unit tests.