Introduce evaluable.Monomial
This patch introduces evaluable.Monomial to represent the sparse tensor
contraction underlying evaluable.factor. This improves the structure of
factored functions that involve arguments of multiple dimensions, which in the
old form had their sparsity shielded after differentiation. As a side effect
the SymProd operation is removed as it is no longer needed.