Reorganize basis specialization
This patch changes the definition of Topology.basis, making it the sole entry
point for basis creation and dispatching to the _tensorial_bases generator.
Additionally, the degree argument is made keyword-only, with a deprecation
warning on postional use, and 'name' argument is renamed (via deprecation) to
'btype' to avoid future clashes with the field name argument, while still
allowing positional use.