Remove psutil and bottombar dependencies
This patch removes the psutil and bottombar modules from pyproject.toml. Having
them as hard dependencies is unreasonable considering that they merely provide
convenience, installing them may be nontrivial, and cli.run may not even be
used.

The cli.run method and _util module are modified to gracefully deal with any
missing functionality, by removing the memory status or bottom bar as a whole.
When the modules are installed the old behaviour is automatically restored.