Change library discovery
This patch modifies the _util.loadlib function to search and load a dynamic
link library by its base name, i.e. without any prefix like lib, suffix like
.so, .dylib, or version number. Compared to ctypes.util.find_library it adds
the functionality of searching the site-package directory in case the library
is not found in the default locations.

The _mkl module is modified to use the new loadlib. The configure_mkl devtool
is removed for being no longer required, along with the NUTILS_MATRIX_MKL_LIB
environment variable.