Rephrase bound replacement error message
This patch changes the error message that is raised when replacing an argument
with a bound array from "replacement functions cannot contain spaces, but
replacement for Argument .. contains space ..'" to "replacement functions
cannot be bound to a space, but replacement for Argument .. is bound to ..".
The new wording is thought to better reflect the situation with spaces.