Introduce function.arguments_for
This patch adds the function.arguments_for function, which can be used to
retrieve array arguments. Unlike that .arguments attribute (which will be
deprecated in the next commit) the returned dictionary maps to fully formed
Argument objects, that can directly be used in operations such as derivative.
Furthermore, the function is nutils-dispatched, so that if can be used even if
the array is wrapped in an SI quantity.