limit cached evaluables to arrays in compile
`evaluable.compile` caches invariant intermediate results for a second call to
the compiled function. The implementation incorrectly assumed that every cached
intermediate evaluates to a `numpy.ndarray`, which is only guaranteed for
`evaluable.Array`s, and makes the evaluated content immutable using
`evaluated.setflags(write=False)`, which fails if `evaluated` is not a
`numpy.ndarray`. This patch fixes the problem by restricting the cache to
instances of `evaluable.Array`.