Centralize locate methods
This patch removes all locate methods except for Topology.locate, which
dispatches the work to _locate and _sample, both of which can be redefined by
the Topology subclass. The centralization makes sure that there is a consistent
function signature and a single docstring, and avoids duplication of pre- and
post-processing common to all topologies.