Support locating points on manifolds
`Topology.locate` requires the coordinate system per element to be smooth,
because it uses a Newton iteration to find the element-local coordinate such
that the given geometry evaluates to the given coordinate. This is problematic
when locating points on periodic domains and a geometry that wraps within an
element. Under certain circumstances we can get around this problem by treating
the domain as a manifold, e.g. a periodic line can be treated as a 1D circle in
2D space, and locating points on the manifold. This patch supports locating
points on manifolds by using a least squares solver for the Newton iteration if
necessary.