Fix graph of evaluables with nested loops
When building a graph of `Evaluable`s, all `Evaluable`s that are to be
evaluated in the same loop are grouped in a subgraph. The sifting logic,
however, contains a flaw. If there are two nested loops, say outer loop `a`
with index `i` and inner loop `b` with index `j`, and some evaluable `d`
depends on index `j` only,

    a = LoopSum, index i
        b = LoopSum, index j
            c = Add
                i
                d = Mul
                    j
                    2

then `d` will be incorrectly positioned outside of the subgraph of `a`. This
descision is made by `Loop._node`, which walks its descendents until an
`Evaluable` without its own loop index is found and draws those outside of the
loop. This patch fixes the bug by including all indices of *nested* loops in
the inside-outside descision making process. In the above example, when loop
`a` encounters loop `b`, the index `j` is added to the list of inside indices
for the descendents of `b` and `d` will correctly be considered being inside.