Add evaluable take slice
`Take` always returns a copy. When taking a slice, e.g. indices `Range(length)`
or `Range(length) + InsertAxis(offset, length)` it is in general much more
efficient to return a sliced view. This patch adds `_TakeSlice`, which compiles
roughly to `array[...,offset:offset+length]`, and adds a simplification to
`Take` if the indices are of the form `Range(length)` or `Range(length) +
InsertAxis(offset, length)`.