Fix reproduceability issue in Add._inflations
This patch replaces the use of set operations with that of dicts and cleans up
surrounding code. The change fixes random behaviour caused by the fact that
sets are ordered by the hash of their objects, which in this case was the
object id, which is not stable. Dictionaries are stable because they preserve
insertion order.