Hash constants in compiled evaluables
Constants are of compiled evaluables are stored in the compiled function's
globals under `c{index}`. When compiling `Constant(0)` and `Constant(1)` the
source codes are the same. The Nutils hash of compiled evaluables is the hash
of the source code. This means that two different compiled evaluables have the
same Nutils hash. To fix this, this patch stores constants under
`c{nutil_hash(value)}`.