Require function.Custom to be hashable in v10
`Evaluable`s using their `evalf`s via the `e{index}` instance in compiled
functions are not represented in the hash of the compiled functions. To fix
this, all evaluables must compile to `SomeEvaluable.evalf` or use hashable
constants. `function.Custom` can't use the former, so this patch makes
`function.Custom.evalf,partial_derivative` hashable using the name of the class
as identifier if they are not hashable yet and makes hashability mandatory as
of Nutils 10. A follow-up patch reimplements the compile step.