Support explicit spaces in function.grad etc
The `function.grad` family computes gradients of the geometry to the root
coordinate systems of all spaces where the geometry is bound to. Under certain
circumstance, e.g. space-time tensor domain with a geometry that rotates in
time, the gradient w.r.t. a geometry must be computed w.r.t. the space root
coordinates only, not in space and time (which fails to lower anyhow, because
the dimension of space and time is larger than the dimension of the geometry).
This patch adds support for specifying the spaces in which the gradient must be
computed.