Add maxretry parameter to System.step
This patch adds the maxretry argument to solver.System.step to limit the number
of timestep bisections that will be performed on error until the function bails
out. The default value is set to 2.

It also allows step to be called without timearg or timesteparg arguments,
reducing step's functionality to copying arguments to to their prefixed
counterparts.