Make function.factor a function
This patch turns function.factor into a function object in order to make it
picklable.

The issue is caused by the @nutils_dispatch decorator: when the factor class
reduces itself to its class and arguments, pickle complains that the class
object does not match the object it finds in the location of its qualname,
since that is occupied by the decorated variant.