Introduce legacy argument for eval, integrate
This patch adds an optional 'legacy' argument to function.Array.eval,
topology.Topology.integrate, and sample.Sample.integrate. When set to True it
evaluates 2D arrays as matrices and 3D and up to sparse arrays; when set to
False it will evaluate all arrays dense. When left undefined it will issue a
deprecation warning to notify that the legacy mode will disappear in Nutils 10.