Make loop id unique in nested samples
In a follow-up commit the restriction that `LowerArgs` cannot update existing
spaces is lifted, enabling nested integrals. Currently, integrals create a
unique loop id based on the spaces of the integral. With nested integrals, the
loop id is not unique anymore. Usually this is not a problem, but if we're
integrating the same space twice and the number of integration points depends
on the loop id (element index), then we can't distinguish the outer loop id
from the inner loop id anymore. This patch makes the loop id more unique by
adding the dimension of the points to the id. The following example illustrates
how the assignment of loop ids:

    # points shape: (), spaces: ()
    Integral X, 0d points -> loop id: _sample_X_0
        # points shape: (), spaces: (X,)
        Integral X, 1d points dependent on loop index -> loop id: _sample_X_0
            # points shape: (Arg(_sample_X_0),), spaces: (X,)
            Integral X, 0d points -> loop id: _sample_X_1
                # points_shape: (Arg(_Sample_X_0),), spaces: (X,)