Add System.construct/deconstruct, add assemble arg
This patch adds the System helper methods .construct and .deconstruct, and
removes the prepare_solution_vector method. It also adds an optional second
argument to System.assemble,

The deconstruct method takes the arguments and constraints dictionaries and
returns an argument-like dictionary and a numpy array. The former contains all
the known values, i.e. all the non-trial arguments and all the constrained
values, with unknown trial values set to nan. The latter is a 1D array
containing the initial guess of the flattened and concatenated unknowns.

The construct method is the reverse operation, constructing the arguments
dictionary from the arguments-like dictionary and a solution vector.

When the System.assemble method is provided a second argument, the first is
assumed to be argument-like and the pair used in argument construction upon
entry. Upon return, the jacobian and residual are restricted to the free dofs,
which allows solvers to operate directly on the relevant solution space.