Remove System.is_constant
This patch removes the System.is_constant attribute, which was used to signal
that either 1. the problem is asymmetric and the residual is constant or 2. the
problem is symmetric and the value is constant. Both situations commonly occur
as a result of replacing the trial arguments linear problems by zero. However,
it is not necessary to explicitly cache values during assembly, as constant
elements are automatically cached interior to the compiled function. The only
relevant object to hold on to is the assembled sparse matrix, which may contain
matrix factors. Otherwise, the assemble method is simplified considerably by
not accounting for constant vectors or values.