Add LowerArg, restructure LowerArgs
To prepare for allowing nested integrals and samples, which requires spaces in
`LowerArgs` to be repeated, restructure `LowerArgs` such that args for spaces
are not stored in a dict, but as a sequence of quartets: `space`, `transforms`,
`index` and `coordinates`, bundled in the new dataclass `LowerArg` (singular).
For now, repeating spaces is disallowed.