Add solver.Direct, remove yielded value
This patch changes the 'method' parameter for System.solve, which remains a
callable but one that can now either return or yield tuple pairs of the
ArrayDict solution and associated residual norm. The value is no longer
included; instead, it is evaluated separately for logging upon return for
symmetric problems.