Reimpl function.opposite() using ~-prefixed spaces
Currently the `LowerArgs` contain a tuple of transforms sequences for each
space. Only the first item of this tuple is used in general.
`function.opposite()` reverses this tuple and passes the reversed `LowerArgs`
to its argument.

This patch reimplements `function.opposite()`. Instead of passing tuples of
transforms sequences to `LowerArgs`, the opposite transforms sequence must be
provided under `~{space}`. `function.opposite()` then swaps the ~-prefixed
space with the unprefixed space and continues lowering. Besides ditching the
tuple, this has the added benefit of allowing the coordinates and index to
change as well, which enables creating an interface using `Sample.zip`.