Log linear solver message at level debug
This patch removes the "solve" logging context from Matrix.solve to avoid the
nested "solve > solve" context, and reduces the log level of the two messages
"solving .. dof system to machine precision using .. solver" and "solver
returned with residual .." from info to debug to reduce the verbosity of the
solution process. The level of "skipping linear solver because initial vector is
within tolerance" is raised to warning as this may be an indication that a
tolerance is set too low. Info messages are now exclusively emitted by
solver.System, with the text "residual" expanded to "residual norm", and one
info log for residual norm added to solve_constraints.