Use System.solve tol as matrix atol by default
This patch provides the tol argument of System.solve to the automatically
created method as the atol linear solver argument, in case no method was
explicitly provided. The rationale is that linear solvers such as arnoldi or
gmres need not iterate further than the tolerance of the solution process that
they are part of.