Add Sample.rename_spaces()
To support zipping samples of different parts of the same topology, add
`Sample.rename_spaces()`, which returns a sample with spaces renamed according
to a dict. The tools to rename spaces in `function.Array`s, necessary to
evaluate functions on the sample with renamed spaces, will be added in the
future.