Replace _isunique by _ismonotonic
This patch introduces the evaluable._ismonotonic helper function, which tests
its argument for being strict monotonic ascending. The function _isunique is
removed and replaced by _ismonotonic(numpy.sort(arg)). The reason for this is
twofold: 1. it exposes the fact that a uniqueness test is expensive, and 2. it
allows for reuse of the intermediate sorted vector.