Guard against swap-take-inflate in assparse
This patch adds a Guard in Array.assparse to avoid an expensive
SwapTakeInflate, resulting from the Take that unique places on top of the
concatenation of sparse terms. The swap is 1. not necessary because there is no
sparsity to be preserved, and 2. not desired, since the inflations that make up
the concatenation are not a sparse construction, so nothing is gained by
pushing the take up the tree.

Introduction of Guard here speeds up factorization of the cylinderflow example
by 70% (a factor 3.4) with respecto to master, and memory use by 30%.