Support empty block matrix assembly
This patch adds support for constructing an empty array via
matrix.assemble_block_csr, which would fail for trying to concatenate an empty
list. This introduces a new requirement for the block arguments which must now
all be numpy arrays: this is required to carry the dtype which is otherwise
unknown in absense of values.