Make integer arrays 64 bit on all platforms
This patch adds the evaluable.Array.ast_dtype property, which serves two
purposes: 1. removing the repeated _pyast.Variable(self.dtype.__name__) for
every operation with a dtype argument, and 2. specifying the precise bit length
for every data type. In particular this forces all integers to be 64 bit, which
overrides Windows behaviour of defaulting to 32 bit integers.